Estimate a INGARCH model

Estimate Integer-valued Generalized Autoregressive Conditional Heteroscedasticity model with Poisson or Negative Binomial distribution. Also is provide a automatic parameter algorithm selection for the Autorregressive and Moving Avarege params

Usage

INGARCH(
  formula,
  ic = c("aic", "bic", "qic"),
  link = c("identity", "log"),
  distr = c("poisson", "nbinom"),
  algorithm = c("naive_search", "arma_to_ingarch"),
  trace = FALSE
)

Arguments

formula

Model specification (see "Specials" section).

ic

Character, can be 'aic' 'bic' or 'qic'. The information criterion used in selecting the model.

link

Character, can be 'identity' or 'log' The link function used for the generalized model

distr

Character, can be 'poisson' or 'nbinom'. The probabilty distribution used for the generalized model

algorithm

Character, specifies the automatic parameter selection algorithm. Can be 'naive_search' or 'arma_to_ingarch'. If 'naive_search' is selected, a search in a 4x4 matrix parameter space is performed, where the model to minimize the criterion value is selected. If 'arma_to_ingarch' is selected, uses an auto_arma as the starting point for the selection algorithm. The ‘arma_to_ingarch’ is the only one to perform a seasonal adjustment

trace

Logical. If the automatic parameter algorithm is runnig, print the path to the best model estimation

Value

A model specification.

Specials

pq

pq defines the non-seasonal autoregressive and moving avarages terms, it can be define by the user, or if it's omited, the automatic parameter selection algorithm is trigered The automatic parameter selection algorithm gonna fit the best model based on the information criterion

PQ

PQ defines the seasonal autoregressive and moving avarages terms, it can be define by the user, or if it's omited, the automatic parameter selection algorithm is trigered (only for 'arma_to_ingarch' algorithm) The automatic parameter selection algorithm gonna fit the best model based on the information criterion

xreg

Exogenous regressors can be included in a INGARCH model without explicitly using the `xreg()` special. Common exogenous regressor specials as specified in [`common_xregs`] can also be used. These regressors are handled using [stats::model.frame()], and so interactions and other functionality behaves similarly to [stats::lm()].

The inclusion of a constant in the model follows the similar rules to [`stats::lm()`], where including `1` will add a constant and `0` or `-1` will remove the constant. If left out, the inclusion of a constant will be determined by minimising `ic`.

If a xreg is provided, the model forecast is not avaliable

xreg(..., fixed = list())

`...`	Bare expressions for the exogenous regressors (such as `log(x)`)

`fixed`	A named list of fixed parameters for coefficients. The names identify the coefficient, and should match the name of the regressor. For example, `fixed = list(constant = 20)`.

Examples

# \donttest{
# Manual INGARCH specification
tsibbledata::aus_production |>
  fabletools::model(manual_ing = INGARCH(Beer ~ pq(1,1)))
#> # A mable: 1 x 1
#>        manual_ing
#>           <model>
#> 1 <INGARCH(1, 1)>

# Automatic INGARCH specification
 tsibbledata::aus_production |>
fabletools::model(auto_ing_naive =
                    INGARCH(Beer,
                            ic = 'aic',
                            trace = TRUE,
                           algorithm = 'naive_search'),
                  auto_ing_arm_ing =
                    INGARCH(Beer,
                            ic = 'aic',
                            trace = TRUE,
                            algorithm = 'arma_to_ingarch'))
#> INGARCH[0, 0], AIC:5735.90 
#> Warning: Without dependence on past observations the dependence on past values of the
#> linear predictor has no effect. Choose the model wisely.
#> Warning: There is almost no serial dependence estimated in the data. This might be
#> appropriate but could just as likely indicate a problem with the optimisation. Please check
#> results carefully.
#> INGARCH[0, 1], AIC:5737.90 
#> Warning: Without dependence on past observations the dependence on past values of the
#> linear predictor has no effect. Choose the model wisely.
#> Warning: There is almost no serial dependence estimated in the data. This might be
#> appropriate but could just as likely indicate a problem with the optimisation. Please check
#> results carefully.
#> INGARCH[0, 2], AIC:5739.90 
#> Warning: Without dependence on past observations the dependence on past values of the
#> linear predictor has no effect. Choose the model wisely.
#> Warning: There is almost no serial dependence estimated in the data. This might be
#> appropriate but could just as likely indicate a problem with the optimisation. Please check
#> results carefully.
#> INGARCH[0, 3], AIC:5741.90 
#> INGARCH[1, 0], AIC:3744.19 
#> INGARCH[1, 1], AIC:3140.70 
#> INGARCH[1, 2], AIC:3593.77 
#> INGARCH[1, 3], AIC:5712.87 
#> INGARCH[2, 0], AIC:4137.00 
#> INGARCH[2, 1], AIC:5669.15 
#> INGARCH[2, 2], AIC:5693.49 
#> INGARCH[2, 3], AIC:5706.39 
#> INGARCH[3, 0], AIC:5670.90 
#> INGARCH[3, 1], AIC:5653.17 
#> INGARCH[3, 2], AIC:5692.41 
#> INGARCH[3, 3], AIC:5731.43 
#> Warning: 1 error encountered for auto_ing_arm_ing
#> [1] unused arguments (link = "identity", distr = "poisson")
#> # A mable: 1 x 2
#>    auto_ing_naive auto_ing_arm_ing
#>           <model>          <model>
#> 1 <INGARCH(1, 1)>     <NULL model>
                            # }